+++ examples/multi_selection/main.qml
@@ -2,7 +2,6 @@ import QtQuick 2.12
 import QtQuick.Window 2.11
 import QtQuick.Controls 2.12
 import QtQuick.Layouts 1.12
-import BppTableModel 0.1
 import BppTable 0.1
 
 ApplicationWindow {