+++ examples/simple_w_buttons/main.qml
@@ -1,7 +1,6 @@
 import QtQuick 2.12
 import QtQuick.Window 2.11
 import QtQuick.Controls 2.12
-import BppTableModel 0.1
 import BppTable 0.1
 
 Window {